app/requirements.txt
aerich==0.8.1
aiofiles==24.1.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.2.1
click==8.1.8
dictdiffer==0.9.0
h11==0.14.0
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
nexios==1.0.2
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.3.2
python-multipart==0.0.20
pytz==2024.2
sniffio==1.3.1
tomli_w==1.2.0
tortoise-orm==0.23.0
typing_extensions==4.12.2
uvicorn==0.34.0
